refactor: EntityPersister 생성 책임 이관
